arreglo de tests y se crea data base initializer
