+++ backend/src/main/java/com/example/demo/services/StoreService.java
@@ -8,6 +8,7 @@ import com.example.demo.model.store.Store;
 import com.example.demo.model.validator.EntityValidator;
 import com.example.demo.repositories.StoreRepository;
 import com.example.demo.repositories.merchandise.MerchandiseRepository;
+import com.example.demo.repositories.storeSchedule.StoreScheduleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -21,6 +22,8 @@ public class StoreService implements IStoreService {
     @Autowired
     private StoreRepository storeRepository;
 
+    @Autowired
+    private StoreScheduleRepository storeScheduleRepository;
     @Autowired
     private MerchandiseRepository merchandiseRepository;
 
@@ -47,6 +50,7 @@ public class StoreService implements IStoreService {
    @Override
    public Store addStore(Store store) {
        validateStore(store);
+       storeScheduleRepository.save(store.storeSchedule());
        storeRepository.save(store);
        return store;
    }
